requirements.txt
streamlit
torch
torchvision
scikit-learn
plotly
opencv-python
Pillow
numpy
tqdm
matplotlib
huggingface_hub